fix todo (#4225)
* fix todo

+++ native/cocos/core/assets/TextureBase.h
@@ -254,7 +254,7 @@ protected:
     gfx::Sampler *   _gfxSampler{nullptr};
     gfx::Device *    _gfxDevice{nullptr};
 
-    uint64_t _textureHash{0};
+    uint32_t _textureHash{0};
 
 private:
     CC_DISALLOW_COPY_MOVE_ASSIGN(TextureBase);
